Fix undefined steps for partial workflow (#5)
Fix type error for `Error: Cannot read properties of undefined
(reading 'status')`. I believe it's from the `steps[key].status ==
Status.Failed` if the value is in the object but undefined. Removed the
class for an object that is allowed to be partial and added more `?.`
guards for field access.