Fix archive for non-existent labels (#14)
This PR fixes two issues with the archive step. First is a typo in the
default event name of `archive_labels`. This should be
`github.event.ref` not `github.event.ref_name. The other fix is for more
robust handling of multiple non existent labels. For `delete` events we
cannot guarantee the branch or tag is matched to a label in the BSR. We
now iterate over all labels and handle missing failures.