Improve GitHub summary for different events
This changes the summary output to be based on the declared steps from
the input. For the default behaviour on pull requests this will output
the same GitHub summary and comment on the PR. For other events and
different inputs the output will now better reflect all the steps that
were run.