Improve links to example workflows (#13)
This PR adds a new section for linking to the push on changes example,
which was a requested example. I've also re-ordered the README.md to
move the `Customizing Behaviour` section below the `Version` and
`Authentication` sections. I think this reads better for the user, as
the user will generally want to customize the `Version` and
`Authentication` before customizing the defaults, which is a more
advanced use case.

---------

Co-authored-by: Nick Snyder <nsnyder@buf.build>