Fix push with no configured module names (#31)
This PR allows for better push behaviour when unnamed modules are
present. The action will now read the `buf.yaml` config file to check
for missing names before running the push or archive steps. With the
latest release of `buf`@v.1.33.0 we can now also use the flag
`--exclude-unnamed` to avoid failing push on unnamed modules when
multiple modules are present. As a test (and example) see the
`test-push-unnamed` step in CI.

In addition is the introduction of outputs in the `action.yml` by
setting the resolved version of `buf`.