Reconfigure inputs for actions (#37)
This PR updates the buf-action from feedback making large changes to the
inputs for the action and updating the documentation to reflect those
changes. Most users shouldn't see changes in behaviour as the default
remain the same. Each commit mentions the affected input values. The
following inputs have been changed:
- `comment`: has been replaced by `pr_comment`
- `disable_symlinks`: removed
- `breaking_against_config`: removed
- `breaking_limit_to_input_files`: removed
- `push_create`: replaced by `push_disable_create` which defaults to
false
- `push_create_visibility`: removed, always unset with the default of
`private`
- `push_labels`: removed, covered by `--git-metadata` flag
- `push_git_metadata`: removed, always set by default
- `push_source_control_url`: removed, always set
- `archive_labels`: removed, sets the ref by default

`breaking_against` now defaults to an empty string and is settable to
override. This allows the action to use the declared `input` value (as
presumably it's a path in the repository) and then configure the
`breaking_against` to use that value for the `subdir` query param. This
should avoid users having to configure the long expression value
themselves.

---------

Co-authored-by: Nick Snyder <nsnyder@buf.build>