Improve summary comments (#46)
Update summary comments to better display the status of the checks by
including an error count, link to the run, and a last updated time.
Update time is useful to ensure that the run reflects the expected
results. The format check has been updated to convert `diff` output to
GitHub annotations. Table is formatted horizontally to aid readability.
There are also links to the BSR for push and archive events.