Support duplicate action comments on PRs
The action may be run multiple times as requested. This isn't supported by
the PR comment status update as it presumes a single run of a v2 style
workspace. To support this case we now dedupe comments on workflow and
job name, setting the job name in the PR to differentiate runs.